chore: Cinema 클래스 도메인 패키지로 이동
