+++ app/src/test/java/woowacourse/movie/domain/cinema/CinemaTest.kt
@@ -5,24 +5,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import woowacourse.movie.domain.reservation.Movie
 import woowacourse.movie.domain.reservation.Screening
-import woowacourse.movie.domain.reservation.ShowtimePolicy
 import java.time.LocalDate
 import java.time.LocalTime
-
-class Cinema(
-    screenings: List<Screening>,
-    private val showtimePolicy: ShowtimePolicy,
-) {
-    val screenings = screenings.map { it.copy() }
-
-    fun showtimes(
-        screening: Screening,
-        date: LocalDate,
-    ): List<LocalTime> {
-        val showtimes = screening.showtimes(date, showtimePolicy)
-        return showtimes
-    }
-}
 
 class CinemaTest {
     private lateinit var cinema: Cinema
